refactor: api/common 하위의 util, constant 클래스들 패키지 추가
