+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/common/constants/UrlConstant.kt
@@ -1,4 +1,4 @@
-package com.yapp.muckpot.common
+package com.yapp.muckpot.common.constants
 
 const val LOGIN_URL = "/api/v1/users/login"
 const val SIGN_UP_URL = "/api/v1/users"
@@ -6,4 +6,3 @@ const val EMAIL_REQUEST_URL = "/api/v1/emails/request"
 const val EMAIL_VERIFY_URL = "/api/v1/emails/verify"
 const val USER_PROFILE_URL = "/v1/users/profile"
 const val LOGOUT_URL = "/api/v1/users/logout"
-const val ROOT_URL = "/"
