+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/domains/board/controller/dto/MuckpotUpdateRequest.kt
@@ -5,15 +5,15 @@ import com.yapp.muckpot.common.AGE_MAX
 import com.yapp.muckpot.common.AGE_MIN
 import com.yapp.muckpot.common.CHAT_LINK_MAX
 import com.yapp.muckpot.common.CONTENT_MAX
-import com.yapp.muckpot.common.CONTENT_MAX_INVALID
 import com.yapp.muckpot.common.HHmm
-import com.yapp.muckpot.common.LINK_MAX_INVALID
 import com.yapp.muckpot.common.Location
-import com.yapp.muckpot.common.MAX_APPLY_MIN_INVALID
-import com.yapp.muckpot.common.NOT_BLANK_COMMON
 import com.yapp.muckpot.common.TITLE_MAX
-import com.yapp.muckpot.common.TITLE_MAX_INVALID
 import com.yapp.muckpot.common.YYYYMMDD
+import com.yapp.muckpot.common.constants.CONTENT_MAX_INVALID
+import com.yapp.muckpot.common.constants.LINK_MAX_INVALID
+import com.yapp.muckpot.common.constants.MAX_APPLY_MIN_INVALID
+import com.yapp.muckpot.common.constants.NOT_BLANK_COMMON
+import com.yapp.muckpot.common.constants.TITLE_MAX_INVALID
 import com.yapp.muckpot.domains.board.entity.Board
 import com.yapp.muckpot.domains.board.exception.BoardErrorCode
 import com.yapp.muckpot.exception.MuckPotException
