+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/domains/user/controller/UserController.kt
@@ -1,17 +1,17 @@
 package com.yapp.muckpot.domains.user.controller
 
 import com.yapp.muckpot.common.ResponseDto
-import com.yapp.muckpot.common.ResponseEntityUtil
-import com.yapp.muckpot.common.SecurityContextHolderUtil
+import com.yapp.muckpot.common.constants.EMAIL_AUTH_REQ_RESPONSE
+import com.yapp.muckpot.common.constants.EMAIL_AUTH_RESPONSE
+import com.yapp.muckpot.common.constants.LOGIN_RESPONSE
+import com.yapp.muckpot.common.constants.SIGN_UP_RESPONSE
+import com.yapp.muckpot.common.utils.ResponseEntityUtil
+import com.yapp.muckpot.common.utils.SecurityContextHolderUtil
 import com.yapp.muckpot.domains.user.controller.dto.LoginRequest
 import com.yapp.muckpot.domains.user.controller.dto.SendEmailAuthRequest
 import com.yapp.muckpot.domains.user.controller.dto.SignUpRequest
 import com.yapp.muckpot.domains.user.controller.dto.VerifyEmailAuthRequest
 import com.yapp.muckpot.domains.user.service.UserService
-import com.yapp.muckpot.swagger.EMAIL_AUTH_REQ_RESPONSE
-import com.yapp.muckpot.swagger.EMAIL_AUTH_RESPONSE
-import com.yapp.muckpot.swagger.LOGIN_RESPONSE
-import com.yapp.muckpot.swagger.SIGN_UP_RESPONSE
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiResponse
