+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/domains/user/service/JwtService.kt
@@ -5,18 +5,18 @@ import com.auth0.jwt.JWTVerifier
 import com.auth0.jwt.algorithms.Algorithm
 import com.auth0.jwt.interfaces.DecodedJWT
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.yapp.muckpot.common.ACCESS_TOKEN_BASIC_SECONDS
-import com.yapp.muckpot.common.ACCESS_TOKEN_KEEP_SECONDS
-import com.yapp.muckpot.common.ACCESS_TOKEN_KEY
-import com.yapp.muckpot.common.CookieUtil
-import com.yapp.muckpot.common.JWT_LOGOUT_VALUE
 import com.yapp.muckpot.common.MS
-import com.yapp.muckpot.common.REFRESH_TOKEN_BASIC_SECONDS
-import com.yapp.muckpot.common.REFRESH_TOKEN_KEEP_SECONDS
-import com.yapp.muckpot.common.REFRESH_TOKEN_KEY
-import com.yapp.muckpot.common.USER_CLAIM
-import com.yapp.muckpot.common.USER_EMAIL_CLAIM
+import com.yapp.muckpot.common.constants.ACCESS_TOKEN_BASIC_SECONDS
+import com.yapp.muckpot.common.constants.ACCESS_TOKEN_KEEP_SECONDS
+import com.yapp.muckpot.common.constants.ACCESS_TOKEN_KEY
+import com.yapp.muckpot.common.constants.JWT_LOGOUT_VALUE
+import com.yapp.muckpot.common.constants.REFRESH_TOKEN_BASIC_SECONDS
+import com.yapp.muckpot.common.constants.REFRESH_TOKEN_KEEP_SECONDS
+import com.yapp.muckpot.common.constants.REFRESH_TOKEN_KEY
+import com.yapp.muckpot.common.constants.USER_CLAIM
+import com.yapp.muckpot.common.constants.USER_EMAIL_CLAIM
 import com.yapp.muckpot.common.enums.YesNo
+import com.yapp.muckpot.common.utils.CookieUtil
 import com.yapp.muckpot.domains.user.controller.dto.UserResponse
 import com.yapp.muckpot.domains.user.exception.UserErrorCode
 import com.yapp.muckpot.exception.MuckPotException
